change file structure current structure only contains main.rs and lib.rs lib contains tpm and common, which also include crypto in the future
